+++ src/dart_dataclasses/entry_points/entry_point_init.py
@@ -15,9 +15,11 @@ default = '''
 parsing_path = ./lib
 # For output of non-inserted generated code (MUST BE IN LIB FOR AUTOMATIC IMPORTS TO WORK!!!!)
 output_path = ./lib/mydataclasses
+# For insertion of generated tests of code given the right decorators
+testing_path = ./test
 
 [Options]
-# Options: vscode jetbrains other
+# Options: vscode, pound_vscode, jetbrains, other
 preferred_editor = jetbrains
 warning_message = True
 reference_private_methods = False