+++ tests/test_cache/test_classes.dart
@@ -1,0 +1,11 @@
+import './class.dart';
+
+void main(){
+  group('Yess', (){
+    Yess a = Yess(address: Address(), name: 'a');
+    @CreateTests(Yess, a)
+
+    @CreateTestTemplates(Yess, a)
+
+  });
+}